more zip and archive fixes

+++ analyzer/windows/lib/common/zip_utils.py
@@ -34,7 +34,8 @@ def extract_archive(seven_zip_path, archive_path, extract_path, password="infect
     stdoutput, stderr = p.stdout, p.stderr
     log.debug(f"{p.stdout} {p.stderr}")
     if b"Wrong password" in stderr:
-        shutil.rmtree(extract_path, ignore_errors=True)
+        if not Path(extract_path).match("local\\temp"):
+            shutil.rmtree(extract_path, ignore_errors=True)
         log.debug([seven_zip_path, "x", f"-p{password}", "-y", f"-o{extract_path}", archive_path])
         p = subprocess.run(
             [seven_zip_path, "x", f"-p{password}", "-y", f"-o{extract_path}", archive_path],
